Avoid re-downloading metadata files
Remote metadata files report a content-length of 0, so comparing only the
file sizes of the remote and local file would cause those to be
downloaded again. By contrast they do report the correct
mtime/last-modified time. By comparing the remote and local mtime we
can be fairly sure that we avoid downloading them again unless they
actually changed.

Unfortunately this comparison does not work for remote compressed
archives of all files, since they appear to be modified immediately before
download. So in those cases we fall back to the old way of comparing
the file size, which is luckily reported correctly.